Extracted components from the StreamDetails SessionsTable
